various fixes, some formatting changes, etc.
Changed the algorithm for rolling new player stats a little.
Mainly so it would quit stomping on the stack.

Stopped storing player passwords in plain text.  When existing
players log in, their player data file will be auto updated.
This change adds a dependency of libsodium to the project.

Instructions on how to build libsodium can be found at
https://download.libsodium.org/doc/installation/